+++ src/pages/requests/interface.tsx
@@ -14,6 +14,7 @@ import { RequestCard } from "@components/cards/RequestCard";
 import { FilterRequestModal } from "@components/modals/FilterRequestModal";
 import { SelectedFilters } from "@components/cards/SelectedFilters";
 import { ERequestType } from "@ptypes/humanResourcesRequest.types";
+import { capitalizeFullName } from "@utils/string";
 
 import { RequestsNav } from "./config/nav.config";
 import { IRoute, BoardSections, RequestItem } from "./types";
@@ -378,8 +379,8 @@ function RequestsUI(props: RequestsUIProps) {
                           id={id}
                           title={requestTypeTitle}
                           requestDate={requestDate}
-                          employeeName={employeeName}
-                          employeeSurnames={surnames}
+                          employeeName={capitalizeFullName(employeeName)}
+                          employeeSurnames={capitalizeFullName(surnames ?? "")}
                           taskName={taskName}
                           responsible={responsible ?? ""}
                           onclick={() => {
@@ -390,7 +391,7 @@ function RequestsUI(props: RequestsUIProps) {
                                   state: {
                                     requestNumber: id,
                                     requestDate,
-                                    fullStaffName: `${employeeName} ${surnames}`,
+                                    fullStaffName: `${capitalizeFullName(employeeName)} ${capitalizeFullName(surnames ?? "")}`,
                                     title: requestTypeTitle,
                                     status: taskName,
                                     statusOptions,
